requirements/requirements_wsl_gpu_organized.txt
# WSL2 Ubuntu TensorFlow GPU Environment Requirements
# Generated: October 7, 2025
# Environment: WSL2 Ubuntu + TensorFlow 2.20.0 + CUDA 12.0
# Virtual Environment: ~/.venvs/tf

# =============================================================================
# CORE MACHINE LEARNING FRAMEWORK
# =============================================================================
tensorflow==2.20.0
keras==3.11.3
numpy==2.1.3
scipy==1.16.2

# =============================================================================
# NVIDIA CUDA LIBRARIES (CUDA 12.0)
# =============================================================================
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.13.1.26
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.3
nvidia-nvjitlink-cu12==12.9.86

# =============================================================================
# IMAGE PROCESSING AND COMPUTER VISION
# =============================================================================
opencv-python==4.12.0.88
pillow==11.3.0
scikit-image==0.25.2
imageio==2.37.0
tifffile==2025.10.4

# =============================================================================
# TENSORFLOW DEPENDENCIES
# =============================================================================
absl-py==2.3.1
astunparse==1.6.3
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
libclang==18.1.1
ml_dtypes==0.5.3
namex==0.1.0
opt_einsum==3.4.0
optree==0.17.0
protobuf==5.29.5
termcolor==3.1.0
wrapt==1.17.3

# =============================================================================
# VISUALIZATION AND MONITORING
# =============================================================================
tensorboard==2.20.0
tensorboard-data-server==0.7.2
rich==14.1.0

# =============================================================================
# NETWORKING AND HTTP
# =============================================================================
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10

# =============================================================================
# TEXT PROCESSING AND MARKUP
# =============================================================================
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
Pygments==2.19.2

# =============================================================================
# UTILITIES AND COMPATIBILITY
# =============================================================================
six==1.17.0
typing_extensions==4.15.0
lazy_loader==0.4
networkx==3.5

# =============================================================================
# DEVELOPMENT TOOLS
# =============================================================================
setuptools==80.9.0
wheel==0.45.1
packaging==25.0
Werkzeug==3.1.3

# =============================================================================
# INSTALLATION NOTES
# =============================================================================
# To install this environment:
# 1. Create virtual environment: python3 -m venv ~/.venvs/tf
# 2. Activate environment: source ~/.venvs/tf/bin/activate
# 3. Upgrade pip: python -m pip install --upgrade pip setuptools wheel
# 4. Install requirements: pip install -r requirements_wsl_gpu_organized.txt
#
# Alternative single command for TensorFlow GPU:
# pip install 'tensorflow[and-cuda]'
#
# Performance: GPU acceleration provides ~6x speedup over CPU
# Compatible with: CUDA 12.0, cuDNN 9.13.1, Python 3.12
